+++ CHANGELOG.md
@@ -4,6 +4,10 @@
 
 [Unreleased]: https://github.com/chaostoolkit-incubator/chaostoolkit-reliably/compare/0.20.0...HEAD
 
+### Added
+
+* Support for execution state: pause/resume and termination
+
 ## [0.20.0][]
 
 [0.20.0]: https://github.com/chaostoolkit-incubator/chaostoolkit-reliably/compare/0.19.0...0.20.0